# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Chat Engine API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "always returns 200",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/connect/{userId}": {
            "get": {
                "description": "promotes an http request to a websocket connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "opens a connection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/join": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "makes user join a room",
                "parameters": [
                    {
                        "description": "room data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PostJoinRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/leave": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "makes user leave a room",
                "parameters": [
                    {
                        "description": "room data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PostLeaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/room": {
            "post": {
                "description": "opens a new room and joins the specified driver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "opens a new room",
                "parameters": [
                    {
                        "description": "room data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PostRoomRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "api.PostJoinRequest": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "api.PostLeaveRequest": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "api.PostRoomRequest": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}